requirements.txt
numpy==1.20.3
opencv-python==4.1.1.26
paddlepaddle-gpu==2.2.0.post101
pandas==1.1.5
Pillow==7.1.2
